Added some fixes to the tables.* code to always set the priority to a default of 10 when reading from the FPGA. This should then allow the read table requests from the P4 Runtime client to work properly.
The P4 Runtime client will also need to use this default priority of 10 for
any rules that need a priority (i.e. LPM, Ternary, Range) when inserting,
deleting or reading rules in order for this fix to work.

We could make changes to the proto/frontend/src/device_mgr.cpp to overwrite
the priority of the insert/delete/read requests coming from the client but
it would be harder to do this for just NP4 targets.